+++ packages/vm/src/compatibility.rs
@@ -26,7 +26,7 @@ const SUPPORTED_IMPORTS: &[&str] = &[
 /// Other optional exports exist, e.g. "query" and "migrate".
 /// This is unlikely to change much, must be frozen at 1.0 to avoid breaking existing contracts
 const REQUIRED_EXPORTS: &[&str] = &[
-    "cosmwasm_vm_version_4",
+    "interface_version_5",
     "init",
     "handle",
     "allocate",
@@ -164,9 +164,9 @@ mod tests {
     use parity_wasm::elements::Internal;
     use std::iter::FromIterator;
 
-    static CONTRACT_0_6: &[u8] = include_bytes!("../testdata/contract_0.6.wasm");
-    static CONTRACT_0_7: &[u8] = include_bytes!("../testdata/contract_0.7.wasm");
-    static CONTRACT: &[u8] = include_bytes!("../testdata/contract.wasm");
+    static CONTRACT_0_6: &[u8] = include_bytes!("../testdata/hackatom_0.6.wasm");
+    static CONTRACT_0_7: &[u8] = include_bytes!("../testdata/hackatom_0.7.wasm");
+    static CONTRACT: &[u8] = include_bytes!("../testdata/hackatom.wasm");
     static CORRUPTED: &[u8] = include_bytes!("../testdata/corrupted.wasm");
 
     fn default_features() -> HashSet<String> {
@@ -228,7 +228,7 @@ mod tests {
     fn check_wasm_old_contract() {
         match check_wasm(CONTRACT_0_7, &default_features()) {
             Err(VmError::StaticValidationErr { msg, .. }) => assert!(msg.starts_with(
-                "Wasm contract doesn't have required export: \"cosmwasm_vm_version_4\""
+                "Wasm contract doesn't have required export: \"interface_version_5\""
             )),
             Err(e) => panic!("Unexpected error {:?}", e),
             Ok(_) => panic!("This must not succeeed"),
@@ -236,7 +236,7 @@ mod tests {
 
         match check_wasm(CONTRACT_0_6, &default_features()) {
             Err(VmError::StaticValidationErr { msg, .. }) => assert!(msg.starts_with(
-                "Wasm contract doesn't have required export: \"cosmwasm_vm_version_4\""
+                "Wasm contract doesn't have required export: \"interface_version_5\""
             )),
             Err(e) => panic!("Unexpected error {:?}", e),
             Ok(_) => panic!("This must not succeeed"),
@@ -350,7 +350,7 @@ mod tests {
         match check_wasm_exports(&module) {
             Err(VmError::StaticValidationErr { msg, .. }) => {
                 assert!(msg.starts_with(
-                    "Wasm contract doesn't have required export: \"cosmwasm_vm_version_4\""
+                    "Wasm contract doesn't have required export: \"interface_version_5\""
                 ));
             }
             Err(e) => panic!("Unexpected error {:?}", e),
@@ -364,7 +364,7 @@ mod tests {
         match check_wasm_exports(&module) {
             Err(VmError::StaticValidationErr { msg, .. }) => {
                 assert!(msg.starts_with(
-                    "Wasm contract doesn't have required export: \"cosmwasm_vm_version_4\""
+                    "Wasm contract doesn't have required export: \"interface_version_5\""
                 ));
             }
             Err(e) => panic!("Unexpected error {:?}", e),